requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
xgboost
yfinance
ccxt
python-dotenv
